# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Piotr Konkol - Praca in\u017cynierska - Podsumowania z u\u017cyciem LLM",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/jobs/summarize": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create summarize job",
        "operationId": "create_summarize_job_api_v1_jobs_summarize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List summarized URLs (grouped, home page)",
        "operationId": "list_summarized_urls_api_v1_jobs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UrlSummaryListItem"
                  },
                  "title": "Response List Summarized Urls Api V1 Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/list": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List all jobs flat (/jobs page)",
        "operationId": "list_all_jobs_flat_api_v1_jobs_list_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobListItemResponse"
                  },
                  "title": "Response List All Jobs Flat Api V1 Jobs List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/by-url": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get all jobs for a URL",
        "operationId": "get_jobs_for_url_api_v1_jobs_by_url_get",
        "parameters": [
          {
            "name": "source_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Exact source URL",
              "title": "Source Url"
            },
            "description": "Exact source URL"
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "One of supported statuses",
              "title": "Status"
            },
            "description": "One of supported statuses"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobStatusResponse"
                  },
                  "title": "Response Get Jobs For Url Api V1 Jobs By Url Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get job status (polling)",
        "operationId": "get_job_status_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Delete Job",
        "operationId": "delete_job_api_v1_jobs__job_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDeletedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeepevalItem": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "name",
          "score",
          "passed",
          "reason"
        ],
        "title": "DeepevalItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreateRequest": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_provider": {
            "type": "string",
            "title": "Model Provider"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          },
          "summary_mode": {
            "type": "string",
            "enum": [
              "simple",
              "sequential",
              "cascade"
            ],
            "title": "Summary Mode",
            "default": "simple"
          },
          "run_deepeval": {
            "type": "boolean",
            "title": "Run Deepeval",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "model_provider",
          "url"
        ],
        "title": "JobCreateRequest"
      },
      "JobCreatedResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "JobCreatedResponse"
      },
      "JobDeletedResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "deleted",
            "title": "Status"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "job_id"
        ],
        "title": "JobDeletedResponse"
      },
      "JobListItemResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "model_provider": {
            "type": "string",
            "title": "Model Provider"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "summary_mode": {
            "type": "string",
            "enum": [
              "simple",
              "sequential",
              "cascade"
            ],
            "title": "Summary Mode"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "source_url",
          "status",
          "title",
          "summary",
          "model_provider",
          "model_name",
          "summary_mode",
          "updated_at"
        ],
        "title": "JobListItemResponse",
        "description": "Flat per-job entry used by GET /api/v1/jobs/list."
      },
      "JobMetrics": {
        "properties": {
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SummaryStatisticalMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "key_takeaways": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KeyTakeawaysMetrics"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "source",
          "summary",
          "key_takeaways"
        ],
        "title": "JobMetrics",
        "description": "All None at insert time \u2014 metrics are computed asynchronously after the job starts."
      },
      "JobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "model_provider": {
            "type": "string",
            "title": "Model Provider"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "summary_mode": {
            "type": "string",
            "enum": [
              "simple",
              "sequential",
              "cascade"
            ],
            "title": "Summary Mode"
          },
          "summary_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SummaryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "metrics": {
            "$ref": "#/components/schemas/JobMetrics"
          },
          "deepeval_metrics": {
            "items": {
              "$ref": "#/components/schemas/DeepevalItem"
            },
            "type": "array",
            "title": "Deepeval Metrics"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageMetadata"
          },
          "raw_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Metadata"
          },
          "raw_output": {
            "type": "string",
            "title": "Raw Output"
          },
          "input_text": {
            "type": "string",
            "title": "Input Text"
          },
          "prompt_template": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Prompt Template"
          },
          "prompt_params": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Prompt Params"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "source_url",
          "status",
          "model_provider",
          "model_name",
          "summary_mode",
          "summary_data",
          "metrics",
          "deepeval_metrics",
          "usage",
          "raw_metadata",
          "raw_output",
          "input_text",
          "prompt_template",
          "prompt_params",
          "created_at",
          "started_at",
          "finished_at",
          "duration_ms",
          "error"
        ],
        "title": "JobStatusResponse"
      },
      "KeyTakeawaysMetrics": {
        "properties": {
          "bullet_count": {
            "type": "integer",
            "title": "Bullet Count"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          }
        },
        "type": "object",
        "required": [
          "bullet_count",
          "char_count"
        ],
        "title": "KeyTakeawaysMetrics"
      },
...
      "SourceMetrics": {
        "properties": {
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          },
          "flesch_kincaid_grade": {
            "type": "number",
            "title": "Flesch Kincaid Grade"
          },
          "text_standard": {
            "type": "number",
            "title": "Text Standard"
          }
        },
        "type": "object",
        "required": [
          "char_count",
          "flesch_kincaid_grade",
          "text_standard"
        ],
        "title": "SourceMetrics"
      },
      "SummaryResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "key_takeaways": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Takeaways"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          }
        },
        "type": "object",
        "required": [
          "title",
          "summary",
          "key_takeaways",
          "source_url"
        ],
        "title": "SummaryResponse",
        "description": "Structured output the LLM is asked to produce."
      },
      "SummaryStatisticalMetrics": {
        "properties": {
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          },
          "flesch_kincaid_grade": {
            "type": "number",
            "title": "Flesch Kincaid Grade"
          },
          "text_standard": {
            "type": "number",
            "title": "Text Standard"
          },
          "length_ratio": {
            "type": "number",
            "title": "Length Ratio"
          }
        },
        "type": "object",
        "required": [
          "char_count",
          "flesch_kincaid_grade",
          "text_standard",
          "length_ratio"
        ],
        "title": "SummaryStatisticalMetrics"
      },
...
      "UrlSummaryListItem": {
        "properties": {
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "completed_count": {
            "type": "integer",
            "title": "Completed Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          },
          "latest_title": {
            "type": "string",
            "title": "Latest Title"
          },
          "latest_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Latest Updated At"
          }
        },
        "type": "object",
        "required": [
          "source_url",
          "completed_count",
          "failed_count",
          "latest_title",
          "latest_updated_at"
        ],
        "title": "UrlSummaryListItem",
        "description": "One entry per unique source_url for the home page list."
      },
      "UsageMetadata": {
        "properties": {
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens",
            "default": 0
          },
          "thinking_tokens": {
            "type": "integer",
            "title": "Thinking Tokens",
            "default": 0
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "input_tokens",
          "output_tokens",
          "thinking_tokens",
          "total_tokens"
        ],
        "title": "UsageMetadata"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}